+++ tests/Integration/PlasterTemplates/Sampler/CompleteModulePlasterTemplate.Integration.Tests.ps1
@@ -34,6 +34,7 @@ Describe 'Complete Module Plaster Template' {
                 # Folders (relative to module root)
 
                 '.github'
+                '.github/ISSUE_TEMPLATE'
                 '.vscode'
                 'output'
                 'output/RequiredModules'
@@ -62,6 +63,10 @@ Describe 'Complete Module Plaster Template' {
 
                 # Files (relative to module root)
 
+                '.github/ISSUE_TEMPLATE/General.md'
+                '.github/ISSUE_TEMPLATE/Problem_with_resource.md'
+                '.github/ISSUE_TEMPLATE/Resource_proposal.md'
+                '.github/PULL_REQUEST_TEMPLATE.md'
                 '.gitattributes'
                 '.gitignore'
                 '.markdownlint.json'
@@ -117,7 +122,7 @@ Describe 'Complete Module Plaster Template' {
                 Force             = $true
 
                 # Template
-                ModuleType        = 'CompleteModule'
+                ModuleType        = 'CompleteSample'
 
                 # Template properties
                 ModuleName        = $mockModuleName